feat/#196 : storage 모듈 삭제
