+++ app/build.gradle.kts
@@ -44,5 +44,4 @@ dependencies {
 	implementation(projects.data.episode)
 	implementation(projects.data.user)
 	implementation(projects.data.mygroup)
-	implementation(projects.data.storage)
 }
