--- data/storage/build.gradle.kts
@@ -1,14 +1,0 @@
-plugins {
-	alias(libs.plugins.mapisode.data)
-}
-
-android {
-	namespace = "com.boostcamp.mapisode.storage"
-}
-
-dependencies {
-	implementation(projects.core.firebase)
-	implementation(projects.domain.storage)
-	implementation(platform(libs.firebase.bom))
-	implementation(libs.firebase.storage)
-}
