+++ data/user/build.gradle.kts
@@ -11,4 +11,5 @@ dependencies {
 	implementation(projects.core.firebase)
 	implementation(platform(libs.firebase.bom))
 	implementation(libs.firebase.firestore)
+	implementation(libs.firebase.storage)
 }
