--- domain/storage/build.gradle.kts
@@ -1,7 +1,0 @@
-plugins {
-	alias(libs.plugins.mapisode.java.library)
-}
-
-dependencies {
-	implementation(projects.core.model)
-}
