--- domain/storage/src/main/java/com/boostcamp/mapisode/storage/StorageRepository.kt
@@ -1,5 +1,0 @@
-package com.boostcamp.mapisode.storage
-
-interface StorageRepository {
-	suspend fun uploadSingleImageToStorage(imageUri: String, uid: String): String
-}
