+++ domain/user/src/main/java/com/boostcamp/mapisode/user/UserRepository.kt
@@ -9,4 +9,5 @@ interface UserRepository {
 	suspend fun isUserExist(uid: String): Boolean
 
 	suspend fun updateUserNameAndProfileUrl(uid: String, userName: String, profileUrl: String)
+	suspend fun uploadSingleImageToStorage(imageUri: String, uid: String): String
 }
