+++ feature/login/build.gradle.kts
@@ -10,6 +10,5 @@ dependencies {
 	implementation(projects.core.auth)
 	implementation(projects.domain.user)
 	implementation(projects.domain.mygroup)
-	implementation(projects.domain.storage)
 	implementation(libs.bundles.coil)
 }
