+++ feature/mypage/build.gradle.kts
@@ -26,6 +26,5 @@ dependencies {
 	implementation(libs.androidx.browser)
 	implementation(projects.core.auth)
 	implementation(projects.domain.user)
-	implementation(projects.domain.storage)
 	implementation(projects.domain.mygroup)
 }
